add view model per screen
