+++ app/src/main/java/com/bridge/notes/model/CreateDealViewModel.kt
@@ -8,29 +8,27 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
-class TournamentDetailsViewModel(
+class CreateDealViewModel(
     private val repository: DataRepository
 ) : ViewModel() {
 
     private val _isLoading = MutableStateFlow(false)
-    val isLoading: StateFlow<Boolean> = _isLoading.asStateFlow()
-
-    private val _currentTournament = MutableStateFlow<Tournament?>(null)
-    val currentTournament: StateFlow<Tournament?> = _currentTournament.asStateFlow()
-
-    fun loadTournament(tournamentId: Long) {
+    fun createDeal(deal: Deal) {
         viewModelScope.launch {
             _isLoading.value = true
-            val tournament = repository.getTournament(tournamentId)
-            _currentTournament.value = tournament
+            repository.createDeal(deal)
             _isLoading.value = false
         }
     }
 
-    fun deleteTournament(tournament: Tournament) {
-        viewModelScope.launch {
-            repository.deleteTournament(tournament)
+    suspend fun getTournamentName(tournamentId: Long): String? {
+        _isLoading.value = true
+        return try {
+            repository.getTournament(tournamentId)?.name
+        } catch (e: Exception) {
+            null
+        } finally {
+            _isLoading.value = false
         }
     }
-
-} 
+}
