+++ app/src/main/java/com/bridge/notes/model/ViewModelFactory.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModelProvider
 import com.bridge.notes.persistence.database.DatabaseProvider
 import com.bridge.notes.persistence.database.DataRepository
 
-class TournamentViewModelFactory(private val context: Context) : ViewModelProvider.Factory {
+class ViewModelFactory(private val context: Context) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         val database = DatabaseProvider.getDatabase(context)
         val repository = DataRepository(database.tournamentDao())
@@ -20,9 +20,25 @@ class TournamentViewModelFactory(private val context: Context) : ViewModelProvid
                 @Suppress("UNCHECKED_CAST")
                 TournamentDetailsViewModel(repository) as T
             }
-            modelClass.isAssignableFrom(DealViewModel::class.java) -> {
+            modelClass.isAssignableFrom(EditTournamentViewModel::class.java) -> {
                 @Suppress("UNCHECKED_CAST")
-                DealViewModel(repository) as T
+                EditTournamentViewModel(repository) as T
+            }
+            modelClass.isAssignableFrom(CreateTournamentViewModel::class.java) -> {
+                @Suppress("UNCHECKED_CAST")
+                CreateTournamentViewModel(repository) as T
+            }
+            modelClass.isAssignableFrom(CreateDealViewModel::class.java) -> {
+                @Suppress("UNCHECKED_CAST")
+                CreateDealViewModel(repository) as T
+            }
+            modelClass.isAssignableFrom(EditDealViewModel::class.java) -> {
+                @Suppress("UNCHECKED_CAST")
+                EditDealViewModel(repository) as T
+            }
+            modelClass.isAssignableFrom(DealDetailViewModel::class.java) -> {
+                @Suppress("UNCHECKED_CAST")
+                DealDetailViewModel(repository) as T
             }
             else -> throw IllegalArgumentException("Unknown ViewModel class")
         }
