+++ app/src/main/java/com/bridge/notes/ui/screen/CreateDealScreen.kt
@@ -4,7 +4,6 @@ import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.Check
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
@@ -13,11 +12,15 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.filled.Check
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.delay
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.ui.Alignment
+import androidx.compose.ui.input.nestedscroll.nestedScroll
 import com.bridge.notes.R
+import com.bridge.notes.model.CreateDealViewModel
 import com.bridge.notes.model.Deal
-import com.bridge.notes.model.DealViewModel
 
 private const val PASSED_OUT = "PASSED OUT"
 
@@ -26,7 +29,7 @@ private const val PASSED_OUT = "PASSED OUT"
 fun CreateDealScreen(
     tournamentId: Long,
     onNavigateBack: () -> Unit,
-    viewModel: DealViewModel
+    viewModel: CreateDealViewModel
 ) {
     var dealNumber by rememberSaveable { mutableStateOf("") }
     var opponents by rememberSaveable { mutableStateOf("") }
@@ -46,23 +49,45 @@ fun CreateDealScreen(
     val doubles = listOf("-", "X", "XX")
     val declarers = listOf("North", "East", "South", "West")
 
+    val scope = rememberCoroutineScope()
+
+    // Add state tracking for the new deal
+    var createdDealId by rememberSaveable { mutableStateOf<Long?>(null) }
+
     // Change declarer state to initialize with first option
     var declarer by rememberSaveable { mutableStateOf(declarers[0]) }
 
+    // Add effect to monitor deal creation
+    LaunchedEffect(createdDealId) {
+        if (createdDealId != null) {
+            delay(200) // Give time for state to update
+            println("Deal creation completed, navigating back")
+            onNavigateBack()
+        }
+    }
     val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior()
 
+    var tournamentName by remember { mutableStateOf("") }
+
+    LaunchedEffect(tournamentId) {
+        tournamentName = viewModel.getTournamentName(tournamentId) ?: ""
+    }
     Scaffold(
+        modifier = Modifier
+            .fillMaxSize()
+            .nestedScroll(scrollBehavior.nestedScrollConnection),
         topBar = {
             TopAppBar(
                 title = { Text(stringResource(R.string.create_deal)) },
                 navigationIcon = {
                     IconButton(onClick = onNavigateBack) {
                         Icon(
-                            Icons.AutoMirrored.Filled.ArrowBack,
+                            imageVector = Icons.AutoMirrored.Filled.ArrowBack,
                             contentDescription = stringResource(R.string.back)
                         )
                     }
                 },
+                scrollBehavior = scrollBehavior,
                 actions = {
                     IconButton(
                         onClick = {
@@ -76,92 +101,223 @@ fun CreateDealScreen(
                                 score = score,
                                 notes = notes
                             )
-                            viewModel.createDeal(newDeal)
-                            onNavigateBack()
-                        }
+
+                            scope.launch {
+                                println("Creating new deal: ${newDeal.id}")
+                                viewModel.createDeal(newDeal)
+                                createdDealId = newDeal.id
+                            }
+                        },
+                        enabled = true
                     ) {
                         Icon(
-                            Icons.Default.Check,
+                            imageVector = Icons.Default.Check,
                             contentDescription = stringResource(R.string.save)
                         )
                     }
-                },
-                scrollBehavior = scrollBehavior
+                }
             )
         }
-    ) { padding ->
+    ) { innerPadding ->
         Column(
             modifier = Modifier
-                .fillMaxSize()
-                .padding(padding)
+                .padding(innerPadding)
                 .padding(16.dp)
+                .fillMaxSize()
                 .verticalScroll(rememberScrollState())
         ) {
+            // Tournament name (uneditable)
+            OutlinedTextField(
+                value = tournamentName,
+                onValueChange = { },
+                label = { Text(stringResource(R.string.deal_tournament_name_label)) },
+                modifier = Modifier.fillMaxWidth(),
+                enabled = false
+            )
+
             OutlinedTextField(
                 value = dealNumber,
                 onValueChange = { dealNumber = it },
                 label = { Text(stringResource(R.string.deal_number)) },
-                modifier = Modifier.fillMaxWidth()
+                modifier = Modifier.fillMaxWidth(),
             )
-            Spacer(modifier = Modifier.height(8.dp))
+
             OutlinedTextField(
                 value = opponents,
                 onValueChange = { opponents = it },
                 label = { Text(stringResource(R.string.opponents)) },
                 modifier = Modifier.fillMaxWidth()
             )
-            Spacer(modifier = Modifier.height(8.dp))
+
+
             Row(
                 modifier = Modifier.fillMaxWidth(),
                 horizontalArrangement = Arrangement.spacedBy(8.dp)
             ) {
-                OutlinedTextField(
-                    value = contractLevel,
-                    onValueChange = { contractLevel = it },
-                    label = { Text("Level") },
+                var levelExpanded by rememberSaveable { mutableStateOf(false) }
+                ExposedDropdownMenuBox(
+                    expanded = levelExpanded,
+                    onExpandedChange = { levelExpanded = !levelExpanded },
                     modifier = Modifier.weight(1f)
-                )
-                OutlinedTextField(
-                    value = contractSuit,
-                    onValueChange = { contractSuit = it },
-                    label = { Text("Suit") },
+                ) {
+                    OutlinedTextField(
+                        value = contractLevel,
+                        onValueChange = {},
+                        readOnly = true,
+                        label = { Text("Level") },
+                        trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = levelExpanded) },
+                        modifier = Modifier.menuAnchor()
+                    )
+                    ExposedDropdownMenu(
+                        expanded = levelExpanded,
+                        onDismissRequest = { levelExpanded = false }
+                    ) {
+                        levels.forEach { level ->
+                            DropdownMenuItem(
+                                text = { Text(level) },
+                                onClick = {
+                                    contractLevel = level
+                                    contract = if (level == PASSED_OUT) {
+                                        PASSED_OUT
+                                    } else {
+                                        "$level$contractSuit $contractDouble"
+                                    }
+                                    levelExpanded = false
+                                }
+                            )
+                        }
+                    }
+                }
+
+                var suitExpanded by rememberSaveable { mutableStateOf(false) }
+                ExposedDropdownMenuBox(
+                    expanded = suitExpanded,
+                    onExpandedChange = {
+                        if (contractLevel != PASSED_OUT) suitExpanded = !suitExpanded
+                    },
                     modifier = Modifier.weight(1f)
-                )
-                OutlinedTextField(
-                    value = contractDouble,
-                    onValueChange = { contractDouble = it },
-                    label = { Text("Double") },
+                ) {
+                    OutlinedTextField(
+                        value = contractSuit,
+                        onValueChange = {},
+                        readOnly = true,
+                        label = { Text("Suit") },
+                        enabled = contractLevel != PASSED_OUT,
+                        trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = suitExpanded) },
+                        modifier = Modifier.menuAnchor()
+                    )
+                    ExposedDropdownMenu(
+                        expanded = suitExpanded,
+                        onDismissRequest = { suitExpanded = false }
+                    ) {
+                        suits.forEach { suit ->
+                            DropdownMenuItem(
+                                text = { Text(suit) },
+                                onClick = {
+                                    contractSuit = suit
+                                    if (contractLevel != PASSED_OUT) {
+                                        contract = "$contractLevel$contractSuit $contractDouble"
+                                    }
+                                    suitExpanded = false
+                                }
+                            )
+                        }
+                    }
+                }
+
+                var doubleExpanded by rememberSaveable { mutableStateOf(false) }
+                ExposedDropdownMenuBox(
+                    expanded = doubleExpanded,
+                    onExpandedChange = {
+                        if (contractLevel != PASSED_OUT) doubleExpanded = !doubleExpanded
+                    },
                     modifier = Modifier.weight(1f)
-                )
+                ) {
+                    OutlinedTextField(
+                        value = contractDouble,
+                        onValueChange = {},
+                        readOnly = true,
+                        label = { Text("Double") },
+                        enabled = contractLevel != PASSED_OUT,
+                        trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = doubleExpanded) },
+                        modifier = Modifier.menuAnchor()
+                    )
+                    ExposedDropdownMenu(
+                        expanded = doubleExpanded,
+                        onDismissRequest = { doubleExpanded = false }
+                    ) {
+                        doubles.forEach { double ->
+                            DropdownMenuItem(
+                                text = { Text(double) },
+                                onClick = {
+                                    contractDouble = double
+                                    if (contractLevel != PASSED_OUT) {
+                                        contract = "$contractLevel$contractSuit $double"
+                                    }
+                                    doubleExpanded = false
+                                }
+                            )
+                        }
+                    }
+                }
             }
-            Spacer(modifier = Modifier.height(8.dp))
-            OutlinedTextField(
-                value = declarer,
-                onValueChange = { declarer = it },
-                label = { Text(stringResource(R.string.declarer)) },
+
+            var declarerExpanded by rememberSaveable { mutableStateOf(false) }
+            ExposedDropdownMenuBox(
+                expanded = declarerExpanded,
+                onExpandedChange = {
+                    if (contractLevel != PASSED_OUT) declarerExpanded = !declarerExpanded
+                },
                 modifier = Modifier.fillMaxWidth()
-            )
-            Spacer(modifier = Modifier.height(8.dp))
+            ) {
+                OutlinedTextField(
+                    value = declarer,
+                    onValueChange = {},
+                    readOnly = true,
+                    label = { Text(stringResource(R.string.declarer)) },
+                    enabled = contractLevel != PASSED_OUT,
+                    trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = declarerExpanded) },
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .menuAnchor()
+                )
+                ExposedDropdownMenu(
+                    expanded = declarerExpanded,
+                    onDismissRequest = { declarerExpanded = false }
+                ) {
+                    declarers.forEach { direction ->
+                        DropdownMenuItem(
+                            text = { Text(direction) },
+                            onClick = {
+                                declarer = direction
+                                declarerExpanded = false
+                            }
+                        )
+                    }
+                }
+            }
+
             OutlinedTextField(
                 value = result,
                 onValueChange = { result = it },
                 label = { Text(stringResource(R.string.result)) },
                 modifier = Modifier.fillMaxWidth()
             )
-            Spacer(modifier = Modifier.height(8.dp))
+
             OutlinedTextField(
                 value = score,
                 onValueChange = { score = it },
                 label = { Text(stringResource(R.string.score)) },
-                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-                modifier = Modifier.fillMaxWidth()
+                modifier = Modifier.fillMaxWidth(),
+                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number)
             )
-            Spacer(modifier = Modifier.height(8.dp))
+
             OutlinedTextField(
                 value = notes,
                 onValueChange = { notes = it },
                 label = { Text(stringResource(R.string.notes)) },
-                modifier = Modifier.fillMaxWidth()
+                modifier = Modifier.fillMaxWidth(),
+                minLines = 3
             )
         }
     }
